requirements.txt
whisper
gtts
playsound
pyaudio
tk
pandas
pyyaml
openai
torch
scipy
faster-whisper
